# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Native Call Center API",
    "version": "1.0.0",
    "summary": "REST and SSE surface of the AI-native call center.",
...
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Same-origin API mount; the SPA is served by the same process."
    }
  ],
...
  "paths": {
...
    "/api-keys": {
      "get": {
        "operationId": "listAPIKeys",
        "tags": [
          "api-keys"
        ],
        "summary": "All API keys",
        "description": "Keys are how a system authenticates. Each carries its own scopes, so a key issued for one integration cannot do another's work.",
        "security": [
          {
            "cookieSession": [
              "keys:manage"
            ]
          },
          {
            "apiKeyBearer": [
              "keys:manage"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Every key, secrets excluded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "post": {
        "operationId": "createAPIKey",
        "tags": [
          "api-keys"
        ],
        "summary": "Issue an API key",
        "description": "Issues a key and returns its secret **once**. Store it now; it cannot be read back.",
        "security": [
          {
            "cookieSession": [
              "keys:manage"
            ],
            "csrfHeader": []
          },
          {
            "apiKeyBearer": [
              "keys:manage"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIKeyWrite"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The key, and its secret, for the only time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyCreated"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api-keys/{keyId}": {
      "parameters": [
        {
          "name": "keyId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getAPIKey",
        "tags": [
          "api-keys"
        ],
        "summary": "One API key",
        "description": "Never carries the secret or its digest.",
        "security": [
          {
            "cookieSession": [
              "keys:manage"
            ]
          },
          {
            "apiKeyBearer": [
              "keys:manage"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "The key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKey"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "patch": {
        "operationId": "updateAPIKey",
        "tags": [
          "api-keys"
        ],
        "summary": "Rename a key or change its scopes",
        "description": "Revocation is not here — it is terminal, so it has an operation of its own. A revoked key cannot be edited.",
        "security": [
          {
            "cookieSession": [
              "keys:manage"
            ],
            "csrfHeader": []
          },
          {
            "apiKeyBearer": [
              "keys:manage"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIKeyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The key as it now stands.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKey"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api-keys/{keyId}/revoke": {
      "parameters": [
        {
          "name": "keyId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "operationId": "revokeAPIKey",
        "tags": [
          "api-keys"
        ],
        "summary": "Revoke a key",
        "description": "Terminal and irreversible: the key stops authenticating immediately and can never be enabled again. Revoking an already revoked key is answered 409 rather than pretended to succeed — an operator who runs this twice should learn that the first one worked. The row stays, because the audit trail names it.",
        "security": [
          {
            "cookieSession": [
              "keys:manage"
            ],
            "csrfHeader": []
          },
          {
            "apiKeyBearer": [
              "keys:manage"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "The revoked key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKey"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "aicc_session",
        "description": "Session cookie issued by POST /auth/login, for a person signed in at a browser. The cookie name is configurable (AICC_SESSION_COOKIE); aicc_session is the default. HttpOnly, SameSite=Lax. The scopes it carries are derived from the account's role at sign-in; the role itself never appears on the wire."
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-AICC-Csrf",
        "description": "Anti-CSRF header, required on every mutating request that authenticates with the session cookie. Any non-empty value passes; requiring a custom header forces a CORS preflight on a cross-origin caller. Missing header: 403 FORBIDDEN. It never accompanies apiKeyBearer — this header protects a credential the browser sends automatically, and a key is sent deliberately."
      },
      "apiKeyBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API key, presented as `Authorization: Bearer <key>`, for a system with no browser. The key is the subject; it is not a person and never becomes one. To act for an agent — a CRM dialling for somebody on the floor — send `X-AICC-Agent-ID: <agentId>` alongside it. Without that header the subject acts for nobody, and an agent-scoped operation answers AGENT_REQUIRED. The header is refused outright on a session request (AGENT_IMPERSONATION_NOT_ALLOWED), whatever the account's role. A key carries exactly the scopes it was issued with; missing one is INSUFFICIENT_SCOPE."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "The request is malformed: unparseable body or an invalid identifier. Codes VALIDATION_FAILED, USER_DATA_TOO_LARGE.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "No session or the session expired. Codes SESSION_EXPIRED, INVALID_CREDENTIALS.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The identity lacks the role, the CSRF header is missing, the account is suspended, or the account is not an agent. Codes FORBIDDEN, USER_SUSPENDED, NOT_CALL_PARTY.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "The target does not exist. Codes NOT_FOUND, CALL_NOT_FOUND.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Conflict": {
        "description": "The request collides with current state. Codes CONFLICT, EXTENSION_IN_USE, EXTENSION_ASSIGNED_TO_AGENT, EXTENSION_POOL_EXHAUSTED, LAST_ADMIN, AGENT_ALREADY_LOGGED_IN, AGENT_NOT_LOGGED_IN, AGENT_NOT_IN_WRAP_UP.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "The body parsed but a field is missing or invalid. Code VALIDATION_FAILED; params.field names the offender when known.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "ServiceUnavailable": {
        "description": "A dependency is unreachable. Codes STORAGE_DOWN, SWITCH_DOWN.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorCode": {
        "type": "string",
        "description": "Machine-readable, translatable failure identifier. The frontend renders errors.<CODE>; the backend never localizes.",
        "enum": [
          "INVALID_CREDENTIALS",
          "SESSION_EXPIRED",
          "FORBIDDEN",
          "AGENT_REQUIRED",
          "AGENT_IMPERSONATION_NOT_ALLOWED",
          "INSUFFICIENT_SCOPE",
          "VALIDATION_FAILED",
          "USER_DATA_TOO_LARGE",
          "NOT_FOUND",
          "METHOD_NOT_ALLOWED",
          "CONFLICT",
          "EXTENSION_IN_USE",
          "EXTENSION_ASSIGNED_TO_AGENT",
          "LAST_ADMIN",
          "EXTENSION_POOL_EXHAUSTED",
          "AGENT_ALREADY_LOGGED_IN",
          "AGENT_NOT_LOGGED_IN",
          "AGENT_NOT_IN_WRAP_UP",
          "CALL_NOT_FOUND",
          "NOT_CALL_PARTY",
          "OPERATION_NOT_ALLOWED_FOR_CALL_TYPE",
          "USER_SUSPENDED",
          "SWITCH_DOWN",
          "STORAGE_DOWN",
          "RATE_LIMITED",
          "INTERNAL"
        ]
      },
      "Error": {
        "type": "object",
        "description": "The single error envelope body: an error code plus interpolation params. Message is diagnostic English, never shown to end users.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "params": {
            "type": "object",
            "description": "Interpolation parameters for the localized message, e.g. {\"field\": \"extensionNumber\"}.",
            "additionalProperties": true
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "APIKeyStatus": {
        "type": "string",
        "description": "ENABLED keys authenticate; REVOKED is terminal and irreversible. There is no third state and no hard delete: a key that ever authenticated is named in the audit trail, and a row that can vanish makes that trail unreadable.",
        "enum": [
          "ENABLED",
          "REVOKED"
        ]
      },
      "APIKey": {
        "type": "object",
        "description": "An API key as it can be read back. The secret is returned once, by the create operation, and never again — this shape carries only the prefix, which is what a person recognises a key by.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "What this key is for, in the operator's words."
          },
          "keyPrefix": {
            "type": "string",
            "description": "The leading, non-secret part of the key, for display and for recognising it in a list. It identifies nothing on its own: authentication looks the key up by the SHA-256 of the whole secret."
          },
          "status": {
            "$ref": "#/components/schemas/APIKeyStatus"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Exactly what this key may do. Names come from x-scopes at the root of this document."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastUsedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When this key last authenticated. Null until it has."
          },
          "revokedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "keyPrefix",
          "status",
          "scopes",
          "createdAt"
        ]
      },
      "APIKeyList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/APIKey"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "APIKeyWrite": {
        "type": "object",
        "description": "What a key is created with.",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "scopes": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string"
            },
            "description": "Names from x-scopes. An unknown name is refused rather than ignored: a key silently missing a capability fails later, somewhere else, for a reason nobody can see."
          }
        },
        "required": [
          "name",
          "scopes"
        ]
      },
      "APIKeyUpdate": {
        "type": "object",
        "description": "What a key can be changed to. Status is not here: the only status change is revocation, which is terminal and has its own operation.",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "scopes": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string"
            }
          }
        }
      },
      "APIKeyCreated": {
        "type": "object",
        "description": "The created key, plus its secret. **This is the only time the secret is ever returned.** It is stored as a SHA-256 digest, so nothing — not this API, not the database, not an administrator — can produce it again. A lost key is revoked and reissued.",
        "allOf": [
          {
            "$ref": "#/components/schemas/APIKey"
          },
          {
            "type": "object",
            "properties": {
              "secret": {
                "type": "string",
                "description": "The credential, to be sent as `Authorization: Bearer <secret>`."
              }
            },
            "required": [
              "secret"
            ]
          }
        ]
      }
    }
  }
}